Use pixel size for score display font

+++ src/libpentobi_gui/ScoreDisplay.h
@@ -31,7 +31,7 @@ protected:
     void resizeEvent(QResizeEvent* event);
 
 private:
-    qreal m_fontSize;
+    int m_fontSize;
 
     QFont m_font;
